refactor(plateform): dedicated api service
